requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-extensions==3.1.3
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==3.2
pipreqs==0.4.10
py4j==0.10.9
PyJWT==1.7.1
pyspark==3.1.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
yarg==0.1.9
